just move info_label processing code to sct.metadata module
