+++ scripts/sct_extract_metric.py
@@ -21,6 +21,8 @@ import sys, io, os, glob, time
 import nibabel as nib
 import numpy as np
 
+from spinalcordtoolbox.metadata import read_label_file
+
 import sct_utils as sct
 from sct_image import get_orientation_3d, set_orientation
 from msct_image import Image
@@ -478,65 +480,6 @@ def extract_metric(method, data, labels, indiv_labels_ids, clusters_labels='', a
         fract_vol_per_label[i_label] = np.sum(labels[i_label])
 
     return metric_in_labels, metric_std_in_labels, fract_vol_per_label
-
-
-def read_label_file(path_info_label, file_info_label):
-    """Reads file_info_label (located inside label folder) and returns the information needed."""
-
-    indiv_labels_ids, indiv_labels_names, indiv_labels_files, combined_labels_ids, combined_labels_names, combined_labels_id_groups, clusters_apriori = [], [], [], [], [], [], []
-
-    # file name of info_label.txt
-    fname_label = os.path.join(path_info_label, file_info_label)
-
-    # Read file
-    try:
-        f = io.open(fname_label, "rb")
-    except IOError:
-        sct.printv('\nWARNING: Cannot open ' + fname_label, 1, 'warning')
-        # raise
-    else:
-        # Extract all lines in file.txt
-        lines = [line.decode("utf-8") for line in f.readlines() if line.rstrip()]
-        lines[-1] += ' '  # To fix an error that could occur at the last line (deletion of the last character of the .txt file)
-
-        # Check if the White matter atlas was provided by the user
-        # look at first line
-        # header_lines = [lines[i] for i in range(0, len(lines)) if lines[i][0] == '#']
-        # info_label_title = header_lines[0].split('-')[0].strip()
-        # if '# White matter atlas' not in info_label_title:
-        #     sct.printv("ERROR: Please provide the White matter atlas. According to the file "+fname_label+", you provided the: "+info_label_title, type='error')
-
-        # remove header lines (every line starting with "#")
-        section = ''
-        for line in lines:
-            # update section index
-            if '# Keyword=' in line:
-                section = line.split('Keyword=')[1].split(' ')[0]
-            # record the label according to its section
-            if (section == 'IndivLabels') and (line[0] != '#'):
-                parsed_line = line.split(', ')
-                indiv_labels_ids.append(int(parsed_line[0]))
-                indiv_labels_names.append(parsed_line[1].strip())
-                indiv_labels_files.append(parsed_line[2].strip())
-
-            elif (section == 'CombinedLabels') and (line[0] != '#'):
-                parsed_line = line.split(', ')
-                combined_labels_ids.append(int(parsed_line[0]))
-                combined_labels_names.append(parsed_line[1].strip())
-                combined_labels_id_groups.append(','.join(parsed_line[2:]).strip())
-
-            elif (section == 'MAPLabels') and (line[0] != '#'):
-                parsed_line = line.split(', ')
-                clusters_apriori.append(parsed_line[-1].strip())
-
-        # check if all files listed are present in folder. If not, ERROR.
-        for file in indiv_labels_files:
-            sct.check_file_exist(os.path.join(path_info_label, file))
-
-        # Close file.txt
-        f.close()
-
-        return indiv_labels_ids, indiv_labels_names, indiv_labels_files, combined_labels_ids, combined_labels_names, combined_labels_id_groups, clusters_apriori
 
 
 def get_slices_matching_with_vertebral_levels(metric_data, vertebral_levels, data_vertebral_labeling, verbose=1):
