+++ scripts/sct_label_vertebrae.py
@@ -25,7 +25,7 @@ from sct_maths import mutual_information
 from msct_parser import Parser
 from msct_image import Image
 import sct_utils as sct
-from sct_warp_template import get_file_label
+from spinalcordtoolbox.metadata import get_file_label
 
 # get path of SCT
 path_sct = os.environ.get("SCT_DIR", os.path.dirname(os.path.dirname(__file__)))
